# services/langgraph/main.py
"""
LangGraph Agent Service
Main FastAPI application for agent orchestration with MCP Bridge.
"""

import time
import asyncio
import json
from contextlib import asynccontextmanager
from fastapi import FastAPI, HTTPException, Request
from fastapi.middleware.cors import CORSMiddleware
from fastapi.responses import JSONResponse, StreamingResponse
from pydantic import BaseModel
from typing import Any, Dict, List, Optional
import uuid

from cobalto.core.config import get_settings
from cobalto.core.logging import setup_logging, get_logger
from cobalto.core.metrics import Metrics, record_http_request
from cobalto.core.health import HealthChecker, HealthStatus
from cobalto.agent.supervisor import SupervisorAgent
from cobalto.agent.base_agent import AgentConfig, AgentType

# MCP imports
from cobalto.mcp.server import MCPServer
from cobalto.mcp.registry.tools import get_tool_registry
from cobalto.mcp.registry.resources import get_resource_registry
from cobalto.mcp.registry.prompts import get_prompt_registry

# Setup logging
setup_logging()
logger = get_logger(__name__)

# Settings
settings = get_settings()

# Lifespan context
@asynccontextmanager
async def lifespan(app: FastAPI):
    """Application lifespan."""
    logger.info("langgraph_service_starting", env=settings.app_env)

    # Initialize components
    app.state.metrics = Metrics("langgraph-api")
    app.state.health_checker = HealthChecker("langgraph-api")
    app.state.supervisor = SupervisorAgent()

    # Initialize MCP Server
    app.state.mcp_server = MCPServer(
        name="cobalto-langgraph-mcp",
        version="0.1.0",
    )

    # Import and register MCP tools
    _register_mcp_tools()
    _register_mcp_resources()
    _register_mcp_prompts()

    # SSE session storage
    app.state.mcp_sessions: Dict[str, asyncio.Queue] = {}

    logger.info("langgraph_service_started", mcp_enabled=settings.mcp_server_enabled)

    yield

    logger.info("langgraph_service_shutting_down")


def _register_mcp_tools():
    """Register all MCP tools."""
    from cobalto.mcp.tools import wazuh, opencti, thehive, response
    logger.info("mcp_tools_registered")


def _register_mcp_resources():
    """Register all MCP resources."""
    from cobalto.mcp.resources import soc_resources
    logger.info("mcp_resources_registered")


def _register_mcp_prompts():
    """Register all MCP prompts."""
    from cobalto.mcp.prompts import soc_prompts
    logger.info("mcp_prompts_registered")


# Create FastAPI app
app = FastAPI(
    title="Cobalto LangGraph Agent Service",
    description="Multi-agent AI orchestration for threat detection and response",
    version="0.1.0",
    lifespan=lifespan,
)

# CORS
app.add_middleware(
    CORSMiddleware,
    allow_origins=["*"],
    allow_credentials=True,
    allow_methods=["*"],
    allow_headers=["*"],
)


# Middleware for metrics
@app.middleware("http")
async def metrics_middleware(request: Request, call_next):
    """Track HTTP metrics."""
    start_time = time.time()
    response = await call_next(request)
    duration = time.time() - start_time

    record_http_request(
        "langgraph-api",
        request.method,
        request.url.path,
        response.status_code,
        duration,
    )
    return response


# Request/Response models
class AnalyzeRequest(BaseModel):
    """Request for alert analysis."""
    alert_id: str
    alert: Dict[str, Any]
    context: Optional[Dict[str, Any]] = None


class AnalyzeResponse(BaseModel):
    """Response from alert analysis."""
    alert_id: str
    status: str
    routing: Dict[str, Any]
    duration_ms: float


class AgentRequest(BaseModel):
    """Request to run a specific agent."""
    agent_type: str
    input_data: Dict[str, Any]
    context: Optional[Dict[str, Any]] = None


class AgentResponse(BaseModel):
    """Response from agent execution."""
    agent_id: str
    agent_type: str
    status: str
    output: Dict[str, Any]
    duration_ms: float


# Routes
@app.get("/health")
async def health():
    """Health check endpoint."""
    checker = app.state.health_checker
    health_result = await checker.run_all_checks()

    status_code = 200
    if health_result.status == HealthStatus.DEGRADED:
        status_code = 200
    elif health_result.status == HealthStatus.UNHEALTHY:
        status_code = 503

    return JSONResponse(
        status_code=status_code,
        content=health_result.to_dict(),
    )


@app.get("/metrics")
async def metrics():
    """Prometheus metrics endpoint."""
    metrics_instance = app.state.metrics
    return JSONResponse(
        content=metrics_instance.get_metrics().decode("utf-8"),
        media_type="text/plain",
    )


@app.post("/agent/analyze", response_model=AnalyzeResponse)
async def analyze_alert(request: AnalyzeRequest):
    """Analyze an alert using the supervisor agent."""
    start_time = time.time()

    try:
        supervisor = app.state.supervisor
        result = await supervisor.run({
            "alert_id": request.alert_id,
            "alert": request.alert,
            "context": request.context or {},
        })

        duration_ms = (time.time() - start_time) * 1000

        return AnalyzeResponse(
            alert_id=request.alert_id,
            status="success",
            routing=result.output.get("routing", {}),
            duration_ms=duration_ms,
        )

    except Exception as e:
        logger.exception("analyze_alert_failed", alert_id=request.alert_id, error=str(e))
        raise HTTPException(status_code=500, detail=str(e))


@app.post("/agent/run", response_model=AgentResponse)
async def run_agent(request: AgentRequest):
    """Run a specific agent."""
    start_time = time.time()

    try:
        # Import and instantiate the appropriate agent
        from cobalto.agent.base_agent import AgentType

        agent_type = AgentType(request.agent_type)

        # For now, return a placeholder
        # In production, this would instantiate and run the actual agent
        duration_ms = (time.time() - start_time) * 1000

        return AgentResponse(
            agent_id=f"{request.agent_type}-{uuid.uuid4().hex[:8]}",
            agent_type=request.agent_type,
            status="success",
            output={"message": f"Agent {request.agent_type} executed successfully"},
            duration_ms=duration_ms,
        )

    except Exception as e:
        logger.exception("run_agent_failed", agent_type=request.agent_type, error=str(e))
        raise HTTPException(status_code=500, detail=str(e))


@app.post("/agent/triage")
async def triage_alert(request: AnalyzeRequest):
    """Triage an alert."""
    start_time = time.time()

    try:
        from services.langgraph.agents.triage import TriageAgent

        agent = TriageAgent()
        result = await agent.run({
            "alert_id": request.alert_id,
            "alert": request.alert,
            "context": request.context or {},
        })

        duration_ms = (time.time() - start_time) * 1000

        return {
            "alert_id": request.alert_id,
            "status": "success",
            "result": result.output,
            "duration_ms": duration_ms,
        }

    except Exception as e:
        logger.exception("triage_alert_failed", alert_id=request.alert_id, error=str(e))
        raise HTTPException(status_code=500, detail=str(e))


@app.post("/agent/analyze-deep")
async def analyze_deep(request: AnalyzeRequest):
    """Deep analysis of an alert."""
    start_time = time.time()

    try:
        from services.langgraph.agents.analysis import AnalysisAgent

        agent = AnalysisAgent()
        result = await agent.run({
            "alert_id": request.alert_id,
            "alert": request.alert,
            "context": request.context or {},
        })

        duration_ms = (time.time() - start_time) * 1000

        return {
            "alert_id": request.alert_id,
            "status": "success",
            "result": result.output,
            "duration_ms": duration_ms,
        }

    except Exception as e:
        logger.exception("analyze_deep_failed", alert_id=request.alert_id, error=str(e))
        raise HTTPException(status_code=500, detail=str(e))


@app.post("/agent/threat-intel")
async def threat_intel_lookup(request: AnalyzeRequest):
    """Threat intelligence lookup."""
    start_time = time.time()

    try:
        from services.langgraph.agents.threat_intel import ThreatIntelAgent

        agent = ThreatIntelAgent()
        result = await agent.run({
            "alert_id": request.alert_id,
            "alert": request.alert,
            "context": request.context or {},
        })

        duration_ms = (time.time() - start_time) * 1000

        return {
            "alert_id": request.alert_id,
            "status": "success",
            "result": result.output,
            "duration_ms": duration_ms,
        }

    except Exception as e:
        logger.exception("threat_intel_failed", alert_id=request.alert_id, error=str(e))
        raise HTTPException(status_code=500, detail=str(e))


@app.post("/agent/response")
async def generate_response(request: AnalyzeRequest):
    """Generate response actions."""
    start_time = time.time()

    try:
        from services.langgraph.agents.response import ResponseAgent

        agent = ResponseAgent()
        result = await agent.run({
            "alert_id": request.alert_id,
            "alert": request.alert,
            "context": request.context or {},
        })

        duration_ms = (time.time() - start_time) * 1000

        return {
            "alert_id": request.alert_id,
            "status": "success",
            "result": result.output,
            "duration_ms": duration_ms,
        }

    except Exception as e:
        logger.exception("generate_response_failed", alert_id=request.alert_id, error=str(e))
        raise HTTPException(status_code=500, detail=str(e))


# =============================================================================
# MCP Bridge Endpoints
# =============================================================================

@app.get("/mcp")
async def mcp_info():
    """MCP server information."""
    mcp_server = app.state.mcp_server
    return JSONResponse(content=mcp_server.get_server_info())


@app.get("/mcp/sse")
async def mcp_sse():
    """SSE endpoint for MCP protocol."""
    if not settings.mcp_server_enabled:
        raise HTTPException(status_code=503, detail="MCP server disabled")

    session_id = f"session-{uuid.uuid4().hex[:8]}"
    queue: asyncio.Queue = asyncio.Queue()
    app.state.mcp_sessions[session_id] = queue

    async def event_generator():
        try:
            # Send connection event
            yield f"event: connected\ndata: {json.dumps({'session_id': session_id})}\n\n"

            while True:
                try:
                    message = await asyncio.wait_for(queue.get(), timeout=30)
                    if message is None:
                        break
                    yield f"data: {json.dumps(message)}\n\n"
                except asyncio.TimeoutError:
                    yield ": keepalive\n\n"
        finally:
            app.state.mcp_sessions.pop(session_id, None)

    return StreamingResponse(
        event_generator(),
        media_type="text/event-stream",
        headers={
            "Cache-Control": "no-cache",
            "Connection": "keep-alive",
            "X-Accel-Buffering": "no",
        },
    )


@app.post("/mcp/messages")
async def mcp_messages(request: Request):
    """Handle MCP JSON-RPC messages."""
    if not settings.mcp_server_enabled:
        raise HTTPException(status_code=503, detail="MCP server disabled")

    try:
        body = await request.json()
        session_id = request.query_params.get("session_id")

        mcp_server = app.state.mcp_server
        response = await mcp_server.handle_message(body)

        # Send to SSE stream if session exists
        if session_id and session_id in app.state.mcp_sessions:
            if response:
                await app.state.mcp_sessions[session_id].put(json.loads(response))

        return JSONResponse(
            content=json.loads(response) if response else {"ok": True},
        )

    except Exception as e:
        logger.exception("mcp_message_error", error=str(e))
        return JSONResponse(
            status_code=500,
            content={
                "jsonrpc": "2.0",
                "error": {"code": -32603, "message": str(e)},
                "id": None,
            },
        )


@app.get("/mcp/tools")
async def mcp_list_tools():
    """List available MCP tools."""
    tool_registry = get_tool_registry()
    tools = tool_registry.list_tools()
    return JSONResponse(
        content={"tools": [tool.model_dump() for tool in tools]}
    )


@app.get("/mcp/resources")
async def mcp_list_resources():
    """List available MCP resources."""
    resource_registry = get_resource_registry()
    resources = resource_registry.list_resources()
    templates = resource_registry.list_templates()
    return JSONResponse(
        content={
            "resources": [r.model_dump() for r in resources],
            "templates": [t.model_dump() for t in templates],
        }
    )


@app.get("/mcp/prompts")
async def mcp_list_prompts():
    """List available MCP prompts."""
    prompt_registry = get_prompt_registry()
    prompts = prompt_registry.list_prompts()
    return JSONResponse(
        content={"prompts": [p.model_dump() for p in prompts]}
    )


@app.post("/mcp/tools/{tool_name}/call")
async def mcp_call_tool(tool_name: str, request: Request):
    """Call an MCP tool directly via HTTP."""
    if not settings.mcp_server_enabled:
        raise HTTPException(status_code=503, detail="MCP server disabled")

    try:
        body = await request.json()
        arguments = body.get("arguments", {})

        tool_registry = get_tool_registry()
        result = await tool_registry.call_tool(tool_name, arguments)

        return JSONResponse(content=result.model_dump())

    except Exception as e:
        logger.exception("mcp_tool_call_error", tool=tool_name, error=str(e))
        raise HTTPException(status_code=500, detail=str(e))


@app.get("/mcp/health")
async def mcp_health():
    """MCP health check."""
    tool_registry = get_tool_registry()
    resource_registry = get_resource_registry()
    prompt_registry = get_prompt_registry()

    return JSONResponse(
        content={
            "status": "healthy",
            "tools_count": len(tool_registry.list_tool_names()),
            "resources_count": len(resource_registry.list_uris()),
            "prompts_count": len(prompt_registry.list_prompt_names()),
            "sessions_count": len(app.state.mcp_sessions) if hasattr(app.state, 'mcp_sessions') else 0,
        }
    )


if __name__ == "__main__":
    import uvicorn
    uvicorn.run(app, host="0.0.0.0", port=8000)